functions/requirements.txt
anyio==4.12.0
blinker==1.9.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
firebase_admin==7.1.0
firebase_functions==0.5.0
Flask==3.1.2
flask-cors==6.0.1
functions-framework==3.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-events==0.5.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
MarkupSafe==3.0.3
msgpack==1.1.2
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
starlette==0.50.0
threadpoolctl==3.6.0
tifffile==2025.10.16
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
uvicorn-worker==0.4.0
watchdog==6.0.0
Werkzeug==3.1.4
